CIV-6892: fix nullpointer
